Update method to get Playwright's version in workflow
The code to retrieve Playwright's version in the GitHub action workflow has been updated. Previously, it was using `npm ls` but now it has been modified to use `pnpm ls` with the `--filter web-e2e` option. This change will ensure a more accurate retrieval of the Playwright test version specific to web end-to-end testing.